pyproject.toml
[project]
name = "aapp-api-services"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
  "django==5.1",  # Later versions were not supported by django-admin-sortable2
  "djangorestframework",
  "adrf",
  "psycopg2-binary",
  "uWSGI",
  "gunicorn",
  "uvicorn",
  "drf_spectacular",
  "python-dotenv",
  "django-debug-toolbar",
  "django-redis",
  "dos2unix",
  "whitenoise",
  "pyproj",
  # polygon operations
  "shapely",
  # pdf
  "fpdf2",
  "babel",
  "qrcode",
  # waste calendar
  "ics",
  # modules
  "djangorestframework-simplejwt",
  "django-csp",
  "django-admin-sortable2",
  # image
  "pillow",
  "pillow-heif",
  # construction work ETL
  "aiohttp",
  "tenacity",
  # notifications
  "firebase_admin",
  # azure
  "django-storages[azure]",
  "azure-ad-verify-token",
  "azure-core",
  "azure-identity",
  "azure-monitor-opentelemetry",
  "azure-monitor-opentelemetry-exporter",
  "opentelemetry-instrumentation-httpx",
  "opentelemetry-api==1.38",  # Temp fix: https://github.com/Azure/azure-sdk-for-python/issues/44237
  # Integration and unit testing
  "freezegun",
  "pytest-django",
  "pytest-dotenv",
  "pytest-env",
  "model-bakery",
  "ruff",
  "aioresponses",
  "pytest-asyncio", # anyio pytest plugin?
  "coverage",
  "responses",
  "amsterdam-django-oidc>=0.2.0",
  "respx",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings.base"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
cache_dir = "/tmp/pytest_cache"
markers = []
addopts = "-m 'not integration'"

[tool.ruff]
cache-dir = "/tmp/.cache/ruff"

[tool.ruff.lint]
# Enable flake8-bugbear (`B`) rules, isort (`I`) rules
select = ["E4", "E7", "E9", "F", "B", "I"]
ignore = ["B904", "B017"]
# Avoid trying to fix flake8-bugbear (`B`) violations.
unfixable = ["B"]

[tool.ruff.lint.per-file-ignores]
"**/{settings}/*" = ["F403", "F405"]

[tool.coverage.run]
branch = true
omit = [
  "*/migrations/*",
  "*/settings/*",
  "*/tests/*",
  "*/integration_tests/*",
  "*/urls.py",
  "*/admin.py",
  "*/admin/*",
  "*/apps.py",
  "notification/utils/patch_utils.py"
]

[tool.coverage.report]
fail_under = 91
show_missing = true
skip_covered = true
